% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rapidoc.R
\docType{package}
\name{rapidoc-package}
\alias{rapidoc}
\alias{rapidoc-package}
\title{rapidoc: Generates RapiDoc documentation from an OpenAPI Specification}
\description{
RapiDoc is a collection of HTML, JavaScript, CSS and fonts assets
that generate Redoc documentation from an OpenAPI Specification.
}
\details{
The main purpose of this package is to enable package authors to create
APIs that are compatible with \url{https://mrin9.github.io/RapiDoc/}
and \url{https://www.openapis.org/}.

To learn more about RapiDoc visit: \url{https://mrin9.github.io/RapiDoc/}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/meztez/rapidoc}
  \item Report bugs at \url{https://github.com/meztez/rapidoc/issues}
}

}
\author{
\strong{Maintainer}: Bruno Tremblay \email{openr@neoxone.com}

Authors:
\itemize{
  \item Mrinmoy Majumdar \email{mrin9@yahoo.com} [copyright holder]
}

Other contributors:
\itemize{
  \item Barret Schloerke \email{barret@rstudio.com} (\href{https://orcid.org/0000-0001-9986-114X}{ORCID}) [contributor]
}

}
