% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.R
\name{rapidoc_index}
\alias{rapidoc_index}
\title{Path to Rapidoc Index}
\usage{
rapidoc_index()
}
\description{
Retrieves the path to the rapidoc index file.
}
\examples{
\dontrun{
if (interactive()) {
  browseURL(rapidoc_index())
} else {
  print(paste("You can use rapidoc under: ", rapidoc_index()))
}
}
}
