% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.R
\name{rapidoc_path}
\alias{rapidoc_path}
\title{Path to RapiDoc Resources}
\usage{
rapidoc_path()
}
\description{
Retrieves the path to rapidoc resources.
}
\examples{
\dontrun{
if (interactive()) {
  browseURL(rapidoc_path())
} else {
  print(paste("You can explore rapidoc resources under: ", rapidoc_path()))
}
}
}
