% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.R
\name{rapidoc_spec}
\alias{rapidoc_spec}
\title{RapiDoc Index File with OpenAPI Path}
\usage{
rapidoc_spec(spec_url = "https://petstore.swagger.io/v2/swagger.json")
}
\arguments{
\item{spec_url}{Url to an openAPI specification}
}
\value{
large string containing the contents of \code{\link{rapidoc_index}()} with
the appropriate speicification path changed to the \code{spec_url} value.
}
\description{
Produces the content for a \code{index.html} file that will attempt to access a
provided OpenAPI Specification URL.
}
\examples{
if (interactive()) {
  rapidoc_spec("https://petstore.swagger.io/v2/swagger.json")
}
}
